Create and move folders into page directory
