+++ src/pages/NotFound/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const NotFound = () => {
+    return (
+        <h1>
+        Not found
+        </h1>
+    );
+};
+
+NotFound.propTypes = {
+
+};
+
+export default NotFound;
