+++ src/tests/ProtextedRoute.test.js
@@ -2,8 +2,8 @@ import ProtectedRoute from '../components/ProtectedRoute';
 import {UserSessionContext} from '../utils/UserContext';
 import {cleanup, render, screen} from '@testing-library/react';
 import {MemoryRouter, Route} from 'react-router-dom';
-import Login from '../Login'
-import Home from '../Home'
+import Login from '../pages/Login'
+import Home from '../pages/Home'
 
 afterEach(cleanup)
